Add files via upload
All submissions for this problem are available.
Elliot has a habit of saving all his data in an encrypted format! Good for you, Elliot! But this time his computer has ran into some problem, and he needs your help.

He usually saves all his data in a key-value pair type. For example, if the world is ellioooottt, it will be saved as eil2o4t3. That is an alphabet and a number if the alphabet is repeated or just the alphabet if it is not repeated.Also the final encrypted string will have alphabetically ordered characters. So given a list of words, encrypt it all for him!


INPUT
The first line of input contains one integer T denoting the number of test cases. The next T lines contains string S which is to be encrypted. All the letters will be lowercase!


OUTPUT
For each test case output the encrypted string in a new line. 
CONSTRAINTS
1< T < 50
10 < S <50

EXAMPLE
Input:
2
ellioooottt
yyoouuu

Output:
eil2o4t3
o2u3y2